// src/store/modules/product/action-types.ts
export enum ProductActionTypes {
  // 产品列表
  ACTION_PRODUCT_LIST = 'ACTION_PRODUCT_LIST',
  ACTION_PRODUCT_SET = 'ACTION_PRODUCT_SET',
  ACTION_PRODUCT_DEL = 'ACTION_PRODUCT_DEL',
  ACTION_PRODUCT_DETAIL = 'ACTION_PRODUCT_DETAIL',
  ACTION_PRODUCT_UPDATE = 'ACTION_PRODUCT_UPDATE',
  ACTION_PRODUCT_PIC_ADD = 'ACTION_PRODUCT_PIC_ADD',
  ACTION_PRODUCT_TAG_OR_TYPE_ADD = 'ACTION_PRODUCT_TAG_OR_TYPE_ADD',
  
  // 库存
  ACTION_PRODUCT_STOCK_LIST = 'ACTION_PRODUCT_STOCK_LIST',
  ACTION_PRODUCT_STOCK_SET = 'ACTION_PRODUCT_STOCK_SET',

  // 产品标签
  ACTION_PRODUCT_TAG_LIST = 'ACTION_PRODUCT_TAG_LIST',
  ACTION_PRODUCT_TAG_SET = 'ACTION_PRODUCT_TAG_SET',
  ACTION_PRODUCT_TAG_DEL = 'ACTION_PRODUCT_TAG_DEL',

  // 产品分类
  ACTION_PRODUCT_CATE_LIST = 'ACTION_PRODUCT_CATE_LIST',
  ACTION_PRODUCT_CATE_SET = 'ACTION_PRODUCT_CATE_SET',
  ACTION_PRODUCT_CATE_DEL = 'ACTION_PRODUCT_CATE_DEL',

  ACTION_PRODUCT_RESOURCE_ALL = 'ACTION_PRODUCT_RESOURCE_ALL',
  
  // 产品关联标签
  ACTION_PRODUCT_RELATIVE_TAG = 'ACTION_PRODUCT_RELATIVE_TAG',
  ACTION_PRODUCT_RELIEVE_TAG = 'ACTION_PRODUCT_RELIEVE_TAG',
  
}